+++ src/Entity/Orders/AdjustmentOrderRepository.php
@@ -2,7 +2,7 @@
 
 namespace App\Entity\Orders;
 
-use App\Entity\OrderRepository;
+use App\Repository\OrderRepository;
 use Doctrine\ORM\QueryBuilder;
 
 class AdjustmentOrderRepository extends OrderRepository
