+++ src/Entity/Orders/MerchandiseOrderRepository.php
@@ -2,7 +2,7 @@
 
 namespace App\Entity\Orders;
 
-use App\Entity\OrderRepository;
+use App\Repository\OrderRepository;
 use Doctrine\ORM\QueryBuilder;
 
 class MerchandiseOrderRepository extends OrderRepository
