+++ src/Entity/Orders/PartnerOrderRepository.php
@@ -2,7 +2,7 @@
 
 namespace App\Entity\Orders;
 
-use App\Entity\OrderRepository;
+use App\Repository\OrderRepository;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\ORM\Tools\Pagination\Paginator;
 use Symfony\Component\HttpFoundation\ParameterBag;
